Widen the index counts, and stop writing live pointers into the file
One format change, so INDEX_HDR goes PFIJ -> PFIK and every device
rebuilds its index once.

The counts go to int32_t, not uint32_t. Unsigned would have been a bad
bug: the builder writes its "untagged" entries backwards from the start
of each array, indexing with -artist_ct and -album_ct, and as uint16_t
those promote to int and negate correctly. Unsigned 32-bit, and the
negation wraps to about four billion and indexes off the end of the
buffer -- on the build path, every time. int32_t raises the ceiling from
65535 to 2^31 and leaves every expression meaning what it meant. Said at
the declaration, because the next person to tidy a count to unsigned
needs to know. It is also `long` on this toolchain, which one debug_log
format string had to be told about.

struct db_summary_t is written whole, and five of its members are
pointers -- this boot's addresses for memory that will be elsewhere, or
gone, next time. They are zeroed in the copy that goes to disk, and the
body writes take their pointers from pfi instead, which is the part that
could have gone wrong: null the struct first and the four write_block
calls after it write from NULL.

Reading back, load_album_index rebinds every pointer as it goes, so a
missed rebind is now a NULL fault rather than a stale address that looks
plausible. load_artist_index loads no album data and now says so three
ways -- count, length and both pointers -- where the count alone was the
whole guard.

+++ apps-ipod/screens/covers/carousel.h
@@ -118,8 +118,15 @@ struct artist_data {
 
 struct db_summary_t {
     uint32_t            header; /*INDEX_HDR*/
-    uint16_t            artist_ct;
-    uint16_t            album_ct;
+    /* Signed, not unsigned, and that is load-bearing. The builder writes the
+     * "untagged" entries backwards from the start of each array, indexing with
+     * -artist_ct / -album_ct (db_summary.c, write_artist_index()). As uint16_t
+     * these promoted to int and negated correctly; as an unsigned 32-bit type
+     * the negation would wrap to about 4 billion and index off the end of the
+     * buffer. int32_t raises the ceiling from 65535 to 2^31 and leaves every
+     * existing expression meaning what it meant. */
+    int32_t             artist_ct;
+    int32_t             album_ct;
     /* What the database looked like when this index was built. Both come from
      * tagcache_get_marks(): commitid moves when tracks are added, serial once
      * per logged play. They say what has happened since, so a later build can
